DataService/requirements.txt
numpy==1.26.3
pillow==10.2.0
tensorflow==2.15.0
flask==3.0.0
SpeechRecognition==3.10.1
requests==2.31.0
g4f==0.2.0.3